requirements.txt
absl-py==0.11.0
appdirs==1.4.4
asgiref==3.3.1
astunparse==1.6.3
beautifulsoup4==4.9.3
black==20.8b1
bs4==0.0.1
cachetools==4.2.0
certifi==2020.12.5
chardet==3.0.4
click==7.1.2
colorama==0.4.3
Django==3.1.4
django-cors-headers==3.5.0
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
gast==0.3.3
google-auth==1.24.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
googletrans==3.1.0a0
grpcio==1.34.0
h11==0.9.0
h2==3.2.0
h5py==2.10.0
hpack==3.0.0
hstspreload==2020.12.22
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
importlib-metadata==3.3.0
joblib==1.0.0
Keras-Preprocessing==1.1.2
lxml==4.6.2
Markdown==3.3.3
mypy-extensions==0.4.3
nltk==3.5
numpy==1.18.5
oauthlib==3.1.0
opt-einsum==3.3.0
pandas==1.1.5
pathspec==0.8.1
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyJWT==1.7.1
python-dateutil==2.8.1
pytz==2020.4
regex==2020.11.13
requests==2.25.1
requests-oauthlib==1.3.0
rfc3986==1.4.0
rsa==4.6
scikit-learn==0.23.2
scipy==1.5.4
six==1.15.0
sniffio==1.2.0
soupsieve==2.1
sqlparse==0.4.1
tensorboard==2.4.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.3.1
tensorflow-estimator==2.3.0
termcolor==1.1.0
threadpoolctl==2.1.0
toml==0.10.2
tqdm==4.54.1
typed-ast==1.4.1
typing-extensions==3.7.4.3
urllib3==1.26.2
Werkzeug==1.0.1
wrapt==1.12.1
xmltodict==0.12.0
zipp==3.4.0
